requirements.txt

google-genai==1.12.1
python-dotenv
